Add files via upload
This file is not complete, but I want to get it pushed up to here as a backup before I continue work on it.

Current major need is to rewrite the original sloppy spaghetti list into a smoother OOP system. Previous was just quick n' dirty testing to prove if it was viable or not. It is, so onto actually making it not look awful and I'll upload that soon