Add files via upload
Refactoring the loop code. Previous version was a draft, and this is more in tune with the final project.

Still need to fix the annoying logic issue where the loop is failing to break out of group 0 and group 1. It's that damn '*' char, but it's absolutely crucial here to have 🙃