Add files via upload
Latest and greatest copy of the graph node test.

The functionality of 2-way switch pathing is fully complete and working as expected which is quite awesome!

The functionality for ∞-way switch pathing now needs to be completed. This will be achieved with a for loop, but I am still debating on where that loop should go. I think within the IncramentStepSwitch function since that'll nestle everything together in one clean and accessible spot. Either way, this is a really really good framework that can scale quite easily over any layout of any size.