Add files via upload
This is the latest version that incorporates the correct switching framework that allows for infinite switch throws. However, the current implementation spawns way too many children to handle the various conditions and the logic needs to be altered to stop that. Also the switches are allowing for impossible path connections, not sure what that's about but at least the core work is there.

Overall I'm pretty happy with where it is currently at