Add files via upload
This is a major update that now allows for infinite switch throws and fixed some other issues related to track end tracking leaving me now with only one critical function remaining deficient: TrackConnections and SwitchConnections direction vectors being unaligned.

This is a rather simple issue to overcome as we just need to flip the polarity of the direction vector used by the ghost agents when encountering one of those conditions. This does require manually flagging the condition into each spot where this exist, but that will honestly be good enough for me if it means have a fully working system. Going to wrap that up later tonight