Add files via upload
This is a critical build update - we have further optimized functions to allow for more modular uses, in one critical case, it's a Lite and Full function difference that allows for various components of the program to alter data only at the needed level. This is going to wildy speed up performance in a lot of prior spots where it was dragging and give me more abilities in the future to chill the number of child agents that are spawning when the program is left unchecked.

The first of two direction inverse flaggers is now functional (albeit needs final implementation). Now if a segment of track connects to another segment that is not aligned to the same vector, it will be detected and an inverse direction bool value applied to the pathing agent when it transfers to the misaligned track. Is this the most efficient manner to handle this? Nope! However, this small cost of performance allows for far more modularity when we begin importing JMRI XML data so we'll take it for now.

I will need to take the framework from the connection inverse flagger and now apply it to the switch inverse flagger. Once that is complete, this entire system will be completely module