Add files via upload
This is a massive update - we can now generate direction vector inversion bools at every track connection segment!! This is really good as now when a train interacts with a segment connection that points to another segment utilizing an opposing polarity it will be able to correctly adjust. This flag even works if it runs over multiple sections which is critical in yard scenarios

The same logic now needs to be applied to switches which is going to be slightly more difficult, but they share enough commonalities that we should have a trivial time working on this. Tedious, but not as ground breaking :)