Add files via upload
We finally have a fully working initializer for both track connections and switch connections!!! We can properly throw boolean flags when jumping connections that tell the pathing system to invert the direction vector and stop paths from immediately reversing over themselves. 

This is incredible - this right here represents a fully working algorithm that just needs a few small tweaks to be perfect. I am anticipating a few more uploads to this repository for structure before we begin working on the production implementation of this system. This is seriously so cool!!!!