Add files via upload
This is the latest build of the project, still have one weird issue happening at group 1 index 0 that is failing to allow data to travel onto the group 2 track, but it's still being investigated. I think the logic is actually working as expected, but maybe something in my test data for the track layout is wrong.

Also, this version now has more support for node-to-node connections where the a node connect or parent node does not contain any sub track elements. I am going to refer to the distinction of these nodes as "single-track" and "multi-track" nodes going forward. While I favor the multi-track node over single-track node, the single track is more native to how JMRI sets up their track connections so I might as well have the capacity to mimic it. This will require a bit more development in regards to how vectors work when interacting with a single track node that is also a switch point, but we'll cross that bridge when needed.

At any rate, this project is coming along really well!