Add files via upload
This is the final update for this testing environment that I intend to push. It currently works as expected and has all the necessary frameworks needed for graphically mapping out a model choo choo layout. There still exist a few weird bugs and an issue with spawning too many child nodes, but it functionally works really well and I am quite happy with where it is at